+++ source/IntegrationTests/Behaviours/IntegrationEvents/TestData/ExampleEnergyResultMessageForActor.cs
@@ -15,7 +15,7 @@
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
 using Energinet.DataHub.EDI.Tests.Infrastructure.OutgoingMessages.Asserts;
 
-namespace Energinet.DataHub.EDI.IntegrationTests.Application.OutgoingMessages.TestData;
+namespace Energinet.DataHub.EDI.IntegrationTests.Behaviours.IntegrationEvents.TestData;
 
 public record ExampleEnergyResultMessageForActor(
     string GridArea,
