+++ source/IntegrationTests/Behaviours/IntegrationEvents/TestData/WholesaleResultForAmountPerChargeDescription.cs
@@ -20,7 +20,7 @@ using NodaTime;
 using Period = Energinet.DataHub.EDI.BuildingBlocks.Domain.Models.Period;
 using Resolution = Energinet.DataHub.EDI.BuildingBlocks.Domain.Models.Resolution;
 
-namespace Energinet.DataHub.EDI.IntegrationTests.Application.OutgoingMessages.TestData;
+namespace Energinet.DataHub.EDI.IntegrationTests.Behaviours.IntegrationEvents.TestData;
 
 public class WholesaleResultForAmountPerChargeDescription
     : TestDataDescription
