+++ source/IntegrationTests/Behaviours/IntegrationEvents/TestData/balance_fixing_01-11-2022_01-12-2022_ga_543_per_ga_v1.csv
