+++ source/IntegrationTests/Behaviours/IntegrationEvents/TestData/wholesale_fixing_01-02-2023_28-02-2023_ga_804_amount_per_charge_v1.csv
